requirements.txt
pip
mesop
discoveryengine
google-genai
google-auth
requests